MAINT: rename test instrument
